+++ tests/countries/test_ivory_coast.py
@@ -9,9 +9,10 @@
 #           ryanss <ryanssdev@icloud.com> (c) 2014-2017
 #  Website: https://github.com/vacanza/holidays
 #  License: MIT (see LICENSE file)
+
 from unittest import TestCase
 
-from holidays.countries import IvoryCoast, CI, CIV
+from holidays.countries.ivory_coast import IvoryCoast, CI, CIV
 from tests.common import CommonCountryTests
